quizlet_bot/requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.6
django-emoji==2.2.2
emoji==2.8.0
filelock==3.12.4
frozenlist==1.4.0
gdown==4.7.1
google-api-core==2.12.0
google-api-python-client==2.102.0
google-auth==2.23.2
google-auth-httplib2==0.1.1
googleapis-common-protos==1.60.0
httplib2==0.22.0
idna==3.4
magic-filter==1.0.12
multidict==6.0.4
oauth2client==4.1.3
protobuf==4.24.4
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyDrive==1.3.1
pyparsing==3.1.1
PySocks==1.7.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.6
yarl==1.9.2
